+++ core-ng-test/src/main/java/core/framework/test/search/MockElasticSearch.java
@@ -1,7 +1,7 @@
 package core.framework.test.search;
 
 import core.framework.api.util.StopWatch;
-import core.framework.impl.search.ElasticSearch;
+import core.framework.impl.search.ElasticSearchImpl;
 import org.elasticsearch.client.Client;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,7 +11,7 @@ import java.nio.file.Path;
 /**
  * @author neo
  */
-public class MockElasticSearch extends ElasticSearch {
+public class MockElasticSearch extends ElasticSearchImpl {
     private final Logger logger = LoggerFactory.getLogger(MockElasticSearch.class);
     private final Path dataPath;