+++ core-ng-test/src/test/java/core/framework/test/IntegrationTest.java
@@ -1,0 +1,11 @@
+package core.framework.test;
+
+import org.junit.runner.RunWith;
+
+/**
+ * @author neo
+ */
+@RunWith(IntegrationTestRunner.class)
+@Context(module = TestModule.class)
+public abstract class IntegrationTest {
+}