+++ core-ng/src/main/java/core/framework/api/module/MongoConfig.java
@@ -3,7 +3,6 @@ package core.framework.api.module;
 import core.framework.api.mongo.Mongo;
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.mongo.MongoImpl;
-import core.framework.impl.mongo.MongoOption;
 
 import java.time.Duration;
 
@@ -11,19 +10,20 @@ import java.time.Duration;
  * @author neo
  */
 public final class MongoConfig {
-    private final MongoOption mongo;
+    private final ModuleContext context;
+    private final MongoImpl mongo;
 
     public MongoConfig(ModuleContext context) {
+        this.context = context;
         if (context.beanFactory.registered(Mongo.class, null)) {
             mongo = context.beanFactory.bean(Mongo.class, null);
         } else {
             if (context.isTest()) {
-                mongo = (MongoOption) context.mockFactory.create(Mongo.class);
+                mongo = context.mockFactory.create(MongoImpl.class);
             } else {
-                MongoImpl mongo = new MongoImpl();
+                mongo = new MongoImpl();
                 context.startupHook.add(mongo::initialize);
                 context.shutdownHook.add(mongo::close);
-                this.mongo = mongo;
             }
             context.beanFactory.bind(Mongo.class, null, mongo);
         }
@@ -34,19 +34,27 @@ public final class MongoConfig {
     }
 
     public void poolSize(int minSize, int maxSize) {
-        mongo.poolSize(minSize, maxSize);
+        if (!context.isTest()) {
+            mongo.poolSize(minSize, maxSize);
+        }
     }
 
     public void slowOperationThreshold(Duration threshold) {
-        mongo.slowOperationThreshold(threshold);
+        if (!context.isTest()) {
+            mongo.slowOperationThreshold(threshold);
+        }
     }
 
     public void tooManyRowsReturnedThreshold(int tooManyRowsReturnedThreshold) {
-        mongo.setTooManyRowsReturnedThreshold(tooManyRowsReturnedThreshold);
+        if (!context.isTest()) {
+            mongo.tooManyRowsReturnedThreshold(tooManyRowsReturnedThreshold);
+        }
     }
 
     public void timeout(Duration timeout) {
-        mongo.timeout(timeout);
+        if (!context.isTest()) {
+            mongo.timeout(timeout);
+        }
     }
 
     public <T> void entityClass(Class<T> entityClass) {